+++ profileStorage/spe-mysql/src/eu/threecixty/cache/ProfileCacheManager.java
@@ -25,7 +25,7 @@ public class ProfileCacheManager {
 	private static final String USER_PROFILE_KEY = "userProfile";
 	private static final int TIME_OUT_TO_GET_CACHE = 200; // in millisecond
 	
-	private MemcachedClient memcachedClient;
+	private List<MemcachedClient> memcachedClients;
 	
 	private static final ProfileCacheManager INSTANCE = new ProfileCacheManager();
 	 private static final Logger LOGGER = Logger.getLogger(
@@ -46,14 +46,14 @@ public class ProfileCacheManager {
 	
 	public void put(UserProfile userProfile) {
 		if (DEBUG_MOD) LOGGER.info("Start putting profile in memory");
-		if (memcachedClient == null) return;
 		if (userProfile == null) return;
 		String _3cixtyUid = userProfile.getHasUID();
 		if (_3cixtyUid == null || _3cixtyUid.equals("")) {
 			if (DEBUG_MOD) LOGGER.info("3cixty UID is null or empty");
 			return;
 		}
-		memcachedClient.set(USER_PROFILE_KEY + _3cixtyUid, 0, userProfile);
+		MemcachedClient memcachedClient = MemcachedUtils.getMemcachedClient(memcachedClients, USER_PROFILE_KEY + _3cixtyUid);
+		if (memcachedClient != null) memcachedClient.set(USER_PROFILE_KEY + _3cixtyUid, 0, userProfile);
 		//profileCaches.put(_3cixtyUid, userProfile);
 		if (DEBUG_MOD) LOGGER.info("Profile stored in memory with 3cixty UID = " + _3cixtyUid);
 		Set <ProfileIdentities> pis = userProfile.getHasProfileIdenties();
@@ -70,33 +70,37 @@ public class ProfileCacheManager {
 				}
 			}
 		}
-		memcachedClient.delete(GOOGLE_UID_FRIENDS_KEY + _3cixtyUid);
+		memcachedClient = MemcachedUtils.getMemcachedClient(memcachedClients, GOOGLE_UID_FRIENDS_KEY + _3cixtyUid);
+		if (memcachedClient != null) memcachedClient.delete(GOOGLE_UID_FRIENDS_KEY + _3cixtyUid);
 		//googleUIDsOfFriends.remove(_3cixtyUid);
 	}
 	
 	public void remove(UserProfile userProfile) {
 		if (DEBUG_MOD) LOGGER.info("Start removing profile in memory");
-		if (memcachedClient == null) return;
 		if (userProfile == null) return;
 		String _3cixtyUid = userProfile.getHasUID();
 		if (_3cixtyUid == null || _3cixtyUid.equals("")) {
 			if (DEBUG_MOD) LOGGER.info("3cixty UID is null or empty");
 			return;
 		}
+		MemcachedClient memcachedClient = MemcachedUtils.getMemcachedClient(memcachedClients, USER_PROFILE_KEY + _3cixtyUid);
+		if (memcachedClient == null) return;
 		memcachedClient.delete(USER_PROFILE_KEY + _3cixtyUid);
 		//profileCaches.remove(_3cixtyUid);
 		if (DEBUG_MOD) LOGGER.info("Profile removed from memory with 3cixty UID = " + _3cixtyUid);
 		//googleUIDsOfFriends.remove(_3cixtyUid);
-		memcachedClient.delete(GOOGLE_UID_FRIENDS_KEY + _3cixtyUid);
+		memcachedClient = MemcachedUtils.getMemcachedClient(memcachedClients, GOOGLE_UID_FRIENDS_KEY + _3cixtyUid);
+		if (memcachedClient != null) memcachedClient.delete(GOOGLE_UID_FRIENDS_KEY + _3cixtyUid);
 	}
 	
 	public UserProfile findProfile(String uid, String source) {
 		if (DEBUG_MOD) LOGGER.info("Start finding profile in memory");
-		if (memcachedClient == null) return null;
 		String generatedID = Utils.gen3cixtyUID(uid,
 				SPEConstants.GOOGLE_SOURCE.equals(source) ? UidSource.GOOGLE : UidSource.FACEBOOK);
 		String _3cixtyUID = uidSourceCaches.get(generatedID);
 		UserProfile profile = null;
+		MemcachedClient memcachedClient = MemcachedUtils.getMemcachedClient(memcachedClients, USER_PROFILE_KEY + _3cixtyUID);
+		if (memcachedClient == null) return null;
 		if (_3cixtyUID != null) {
 			Future<Object> f = memcachedClient.asyncGet(USER_PROFILE_KEY + _3cixtyUID);
 			try {
@@ -128,6 +132,7 @@ public class ProfileCacheManager {
 	
 	public UserProfile getProfile(String _3cixtyUID) {
 		if (DEBUG_MOD) LOGGER.info("Checking in the memory for 3cixtyUID = " + _3cixtyUID);
+		MemcachedClient memcachedClient = MemcachedUtils.getMemcachedClient(memcachedClients, USER_PROFILE_KEY + _3cixtyUID);
 		if (memcachedClient == null) return null;
 		UserProfile profile = null;
 		if (_3cixtyUID != null) {
@@ -157,6 +162,7 @@ public class ProfileCacheManager {
 	
 	public void putGoogleUIDsOfFriens(String _3cixtyUID, List <String> googleUIDs) {
 		if (_3cixtyUID == null || googleUIDs == null) return;
+		MemcachedClient memcachedClient = MemcachedUtils.getMemcachedClient(memcachedClients, GOOGLE_UID_FRIENDS_KEY + _3cixtyUID);
 		if (memcachedClient == null) return;
 		memcachedClient.set(GOOGLE_UID_FRIENDS_KEY + _3cixtyUID, 0, googleUIDs);
 		//googleUIDsOfFriends.put(_3cixtyUID, googleUIDs);
@@ -164,6 +170,7 @@ public class ProfileCacheManager {
 	
 	public List <String> getGoogleUIDsOfFriends(String _3cixtyUID) {
 		if (_3cixtyUID == null) return null;
+		MemcachedClient memcachedClient = MemcachedUtils.getMemcachedClient(memcachedClients, GOOGLE_UID_FRIENDS_KEY + _3cixtyUID);
 		if (memcachedClient == null) return null;
 		Future<Object> f = memcachedClient.asyncGet(GOOGLE_UID_FRIENDS_KEY + _3cixtyUID);
 		try {
@@ -188,13 +195,17 @@ public class ProfileCacheManager {
 	}
 	
 	public void stop() {
-		if (memcachedClient != null) memcachedClient.shutdown();
+		if (memcachedClients != null) {
+			for (MemcachedClient client: memcachedClients) {
+				client.shutdown();
+			}
+		}
 	}
 	
 	private ProfileCacheManager() {
 		//profileCaches = new ConcurrentHashMap<String, UserProfile>();
 		uidSourceCaches = new ConcurrentHashMap<String, String>();
 		//googleUIDsOfFriends = new ConcurrentHashMap<String, List<String>>();
-		memcachedClient = MemcachedUtils.createClient();
+		memcachedClients = MemcachedUtils.createClients();
 	}
 }
